Refactor reflecta to remove hard dependency on Servo.h and Wire.h
Not all sketches need servo.h or wire.h. Removing them can prevent
future conflicts. This also allows the use of other servo libraries,
such as PWMServo, SoftwareServo, the Adafruit Servo library, etc.

note: The servo.h integration is currently busted, regardless of this
change.
