+++ examples/StandardReflecta/StandardReflecta.ino
@@ -1,6 +1,4 @@
 #include <Arduino.h>
-#include <Wire.h>
-#include <Servo.h>
 #include <Reflecta.h>
 
 // Very simple Arduino sketch for using reflecta.  Opens a COM port
